+++ plugin.json
@@ -2,9 +2,7 @@
   "name": "Deck Shelves",
   "author": "Jonathan Santos",
   "api_version": 1,
-  "flags": [
-    "debug"
-  ],
+  "flags": [],
   "publish": {
     "tags": [
       "customization",
