+++ src/components/shelf/shelfStyles.ts
@@ -1,0 +1,316 @@
+import { getPreferredSteamDocument } from "../../runtime/steamHost";
+import { discoverNativeCardDimensions, type NativeCardDims } from "../../core/webpackCompat";
+import { logInfo } from "../../runtime/logger";
+import { CARD_W, CARD_ART_H, CARD_GAP } from "./types";
+
+const STYLE_ID = "deck-shelves-row-style";
+
+let cachedCardRadius = "0px";
+let cachedNativeDims: NativeCardDims | null = null;
+const nativeDimsListeners = new Set<() => void>();
+
+let dimsDebounceTimer: ReturnType<typeof setTimeout> | null = null;
+function debouncedNotifyDims() {
+  if (dimsDebounceTimer) clearTimeout(dimsDebounceTimer);
+  dimsDebounceTimer = setTimeout(() => {
+    dimsDebounceTimer = null;
+    nativeDimsListeners.forEach(cb => cb());
+  }, 300);
+}
+
+export function getCachedCardRadius(): string { return cachedCardRadius; }
+export function getCachedNativeDims(): NativeCardDims | null { return cachedNativeDims; }
+export function onNativeDimsChange(cb: () => void): () => void {
+  nativeDimsListeners.add(cb);
+  return () => { nativeDimsListeners.delete(cb); };
+}
+
+function detectNativeCardRadius(): string {
+  try {
+    const doc = getPreferredSteamDocument();
+    if (!doc) return "0px";
+    const panels = doc.querySelectorAll(".Panel.Focusable");
+    for (let i = 0; i < panels.length; i++) {
+      const panel = panels[i] as HTMLElement;
+      const cls = panel.className ?? "";
+      if (cls.indexOf("ds-card") >= 0 || cls.indexOf("ds-row") >= 0) continue;
+      if (!panel.querySelector("img")) continue;
+      try {
+        const r = getComputedStyle(panel).borderRadius;
+        if (r && r !== "0px") return r;
+      } catch { /* skip */ }
+    }
+  } catch {
+    logInfo("HOME", "detectNativeCardRadius failed");
+  }
+  return "0px";
+}
+
+function ensureStyles() {
+  try {
+    const newRadius = detectNativeCardRadius();
+    const radiusChanged = newRadius !== cachedCardRadius;
+    cachedCardRadius = newRadius;
+    const steamDoc = getPreferredSteamDocument();
+    const newDims = discoverNativeCardDimensions(steamDoc) ?? discoverNativeCardDimensions(document);
+    const dimsChanged = newDims !== null && (
+      !cachedNativeDims ||
+      newDims.width !== cachedNativeDims.width ||
+      newDims.height !== cachedNativeDims.height ||
+      newDims.gap !== cachedNativeDims.gap ||
+      newDims.featuredWidth !== cachedNativeDims.featuredWidth ||
+      newDims.featuredHeight !== cachedNativeDims.featuredHeight
+    );
+    if (newDims) cachedNativeDims = newDims;
+    if (dimsChanged) debouncedNotifyDims();
+    const docs = [document, steamDoc];
+    for (const doc of docs) {
+      if (!doc) continue;
+      if (radiusChanged || dimsChanged) {
+        const existing = doc.getElementById(STYLE_ID);
+        if (existing) existing.remove();
+      }
+      if (!doc.getElementById(STYLE_ID)) {
+        const style = doc.createElement("style");
+        style.id = STYLE_ID;
+        style.textContent = buildStylesheet();
+        doc.head.appendChild(style);
+      }
+
+      try {
+        doc.documentElement.style.removeProperty('--ds-native-heading-color');
+        const headings = doc.querySelectorAll('h2[class], h3[class]');
+        for (const h of Array.from(headings)) {
+          const cls = (h as HTMLElement).className || '';
+          if (/_[A-Za-z0-9_-]{5,}/.test(cls)) {
+            const c = getComputedStyle(h as HTMLElement).color;
+            if (!c || c === 'rgb(0, 0, 0)' || c === 'rgba(0, 0, 0, 0)') continue;
+            const m = c.match(/rgb\((\d+),\s*(\d+),\s*(\d+)\)/);
+            if (m) {
+              const [r, g, b] = [parseInt(m[1]), parseInt(m[2]), parseInt(m[3])];
+              const max = Math.max(r, g, b);
+              const sat = max > 0 ? (max - Math.min(r, g, b)) / max : 0;
+              if (sat < 0.25) continue;
+            }
+            doc.documentElement.style.setProperty('--ds-native-heading-color', c);
+            break;
+          }
+        }
+      } catch {
+        logInfo("HOME", "heading color detection failed");
+      }
+    }
+  } catch {
+    logInfo("HOME", "ensureStyles failed");
+  }
+}
+
+function buildStylesheet(): string {
+  return `
+    :root {
+      --ds-card-radius: ${cachedCardRadius};
+      --ds-card-dim: 0.9;
+      --ds-card-bg: rgba(3, 10, 30, 0.92);
+      --ds-native-card-w: ${cachedNativeDims?.width ?? CARD_W}px;
+      --ds-native-card-h: ${cachedNativeDims?.height ?? CARD_ART_H}px;
+      --ds-native-card-gap: ${cachedNativeDims?.gap ?? CARD_GAP}px;
+    }
+    #deck-shelves-home-root { margin-top: -24px !important; }
+    .ds-row-scroll { scrollbar-width: none; -ms-overflow-style: none; }
+    .ds-row-scroll::-webkit-scrollbar { display: none; width: 0; height: 0; }
+    .ds-card {
+      border-radius: var(--ds-card-radius, ${cachedCardRadius}) !important;
+      filter: brightness(var(--ds-card-dim, 0.9));
+      transition: filter 0.4s cubic-bezier(0, 0.73, 0.48, 1);
+      scroll-margin-top: 90px;
+      scroll-margin-bottom: 52px;
+      scroll-margin-inline-end: 2.8vw;
+    }
+    #deck-shelves-home-root .deck-shelves-root:focus,
+    #deck-shelves-home-root .deck-shelves-root.gpfocus,
+    #deck-shelves-home-root .deck-shelves-root.gpfocuswithin,
+    #deck-shelves-home-root .ds-row-scroll:focus,
+    #deck-shelves-home-root .ds-row-scroll.gpfocus,
+    #deck-shelves-home-root .ds-row-scroll.gpfocuswithin,
+    #deck-shelves-home-root .Panel.gpfocus,
+    #deck-shelves-home-root .Focusable.gpfocus,
+    #deck-shelves-home-root [class*="row"].gpfocus {
+      outline: none !important;
+      border: none !important;
+      box-shadow: none !important;
+      animation: none !important;
+    }
+    #deck-shelves-home-root .ds-card:focus,
+    #deck-shelves-home-root .ds-card.gpfocus {
+      outline: none !important;
+      outline-offset: 0px !important;
+      border: none !important;
+      box-shadow: none !important;
+      z-index: 5;
+    }
+    #deck-shelves-home-root .ds-card::after {
+      content: '' !important;
+      position: absolute !important;
+      top: 0 !important;
+      left: 0 !important;
+      right: 0 !important;
+      bottom: auto !important;
+      height: var(--ds-card-art-h, 100%) !important;
+      border-radius: var(--ds-card-radius, ${cachedCardRadius}) !important;
+      pointer-events: none !important;
+      z-index: 4 !important;
+      opacity: 1 !important;
+      transition: none !important;
+      background: none !important;
+      background-image: none !important;
+      animation: none !important;
+      display: inline !important;
+    }
+    #deck-shelves-home-root .ds-card.gpfocus::after,
+    #deck-shelves-home-root .ds-card:focus::after {
+      height: var(--ds-card-art-h, 100%) !important;
+      bottom: auto !important;
+      border-radius: var(--ds-card-radius, ${cachedCardRadius}) !important;
+    }
+    #deck-shelves-home-root .ds-card .ds-card-shimmer { display: none !important; }
+    @keyframes ds-shelf-shimmer {
+      0% { background-position: 0% 0%; opacity: 0; }
+      40% { opacity: 1; }
+      100% { background-position: -200% 0%; opacity: 0; }
+    }
+    @keyframes ds-focus-pulse {
+      0% { opacity: 0; }
+      40% { opacity: 1; }
+      100% { opacity: 0; }
+    }
+    #deck-shelves-home-root .ds-card *:focus { outline: none !important; box-shadow: none !important; }
+    .ds-card-art {
+      position: absolute !important;
+      inset: 1px !important;
+      height: var(--ds-card-art-h, 100%) !important;
+      padding-top: 0 !important;
+      border-radius: var(--ds-card-radius, ${cachedCardRadius});
+    }
+    .ds-card-art img {
+      border-radius: var(--ds-card-radius, ${cachedCardRadius});
+    }
+    .ds-card.gpfocus .ds-card-art,
+    .ds-card:focus .ds-card-art {
+      z-index: 2;
+    }
+    .ds-card .ds-card-label {
+      opacity: 0;
+      transition: opacity .15s ease;
+    }
+    .ds-card.gpfocus .ds-card-label,
+    .ds-card:focus .ds-card-label {
+      opacity: 1;
+    }
+    .ds-card img { transition: opacity .15s ease; }
+    .ds-compat {
+      position: absolute; bottom: 4px; right: 4px;
+      display: var(--ds-compat-display, flex); align-items: center;
+      background: rgba(0,0,0,0.7);
+      border-radius: 20px;
+      padding: 2px;
+      z-index: 3; pointer-events: none;
+      width: 40px; height: 20px;
+      opacity: 0;
+      transition: opacity .15s ease;
+    }
+    .ds-card.gpfocus .ds-compat,
+    .ds-card:focus .ds-compat { opacity: var(--ds-compat-opacity, 1); }
+    .ds-compat svg { flex-shrink: 0; width: 20px; height: 20px; }
+    .ds-compat-deck-icon { color: var(--custom-compat-icons-deck, rgba(255,255,255,0.84)); }
+    .ds-compat-verified .ds-compat-verdict-icon { color: var(--custom-compat-icons-verified, rgb(89, 191, 64)); }
+    .ds-compat-playable .ds-compat-verdict-icon { color: var(--custom-compat-icons-playable, rgb(255, 200, 44)); }
+    .ds-compat-unsupported .ds-compat-verdict-icon { color: var(--custom-compat-icons-unsupported, rgb(220, 222, 223)); }
+    .ds-compat-unknown .ds-compat-verdict-icon { color: var(--custom-compat-icons-unknown, rgba(255,255,255,0.4)); }
+    .ds-shelf-title {
+      color: var(--ds-native-heading-color, inherit);
+      font-size: 22px;
+      font-weight: 700;
+      letter-spacing: 0.5px;
+    }
+    .ds-shelf-collapse-icon {
+      font-size: 14px;
+      opacity: 0.5;
+      transition: transform 0.2s;
+      display: inline-block;
+    }
+    .ds-card-label-name {
+      color: var(--ds-native-heading-color, inherit);
+      font-size: inherit;
+      line-height: 1.2;
+      font-weight: bold;
+      white-space: nowrap;
+      overflow: visible;
+    }
+    .ds-card-status {
+      display: flex;
+      align-items: center;
+      gap: 6px;
+      opacity: 0.7;
+      font-size: 0.75em;
+      line-height: 1.3;
+      font-weight: bold;
+      text-transform: uppercase;
+      margin-top: 4px;
+      white-space: nowrap;
+      overflow: visible;
+    }
+    .ds-card-status-icon {
+      display: inline-flex;
+      align-items: center;
+      justify-content: center;
+      width: 14px;
+      height: 14px;
+      flex-shrink: 0;
+      line-height: 0;
+    }
+    .ds-card-status-play { color: var(--ds-native-heading-color, rgb(89, 191, 64)); }
+    .ds-more-card-text {
+      font-size: 16px;
+      font-weight: 400;
+      line-height: 1.35;
+      text-align: center;
+    }
+    .ds-card-art-placeholder {
+      font-size: 11px;
+      opacity: 0.5;
+      text-align: center;
+      word-break: break-word;
+    }
+    .ds-card.ds-card--featured .ds-card-label { opacity: 1 !important; }
+    .ds-card.ds-card--featured .ds-card-art img { object-position: center top; }
+  `;
+}
+
+// Single global timer for ensureStyles — shared by all DeckRow instances.
+let globalStyleRefCount = 0;
+let globalStyleTimer: ReturnType<typeof setInterval> | null = null;
+let globalResizeHandler: (() => void) | null = null;
+
+export function globalStylesStart() {
+  if (++globalStyleRefCount === 1) {
+    ensureStyles();
+    globalStyleTimer = setInterval(ensureStyles, 3000);
+    globalResizeHandler = () => ensureStyles();
+    window.addEventListener('resize', globalResizeHandler);
+  }
+}
+
+export function globalStylesStop() {
+  if (--globalStyleRefCount <= 0) {
+    globalStyleRefCount = 0;
+    if (globalStyleTimer) { clearInterval(globalStyleTimer); globalStyleTimer = null; }
+    if (globalResizeHandler) { window.removeEventListener('resize', globalResizeHandler); globalResizeHandler = null; }
+  }
+}
+
+export function formatPlaytime(minutes: number | undefined): string | null {
+  if (!minutes || minutes <= 0) return null;
+  if (minutes < 60) return `${minutes}min`;
+  const hours = minutes / 60;
+  return `${hours.toFixed(1).replace(".", ",")} h`;
+}
